refactor: move client credentials grant type features to their own folder
