+++ src/components/auth/Oauth/ClientCredentials/NoWorkspaceOauthClientCredsFlow.tsx
@@ -5,15 +5,16 @@
 import { useState } from 'react';
 import { Connection, GenerateConnectionRequest } from '@generated/api/src';
 
-import {
-  ClientCredentialsContent,
-  WorkspaceClientCredentialsCreds,
-} from 'components/auth/Oauth/WorkspaceEntry/ClientCredentialsContent';
 import { LoadingCentered } from 'components/Loading';
 import { useApiKey } from 'context/ApiKeyContextProvider';
 import { useProject } from 'context/ProjectContextProvider';
 import { api } from 'services/api';
 import { handleServerError } from 'src/utils/handleServerError';
+
+import {
+  ClientCredentialsCreds,
+  NoWorkspaceClientCredentialsContent,
+} from './NoWorkspaceClientCredentialsContent';
 
 interface NoWorkspaceOauthClientCredsFlowProps {
   provider: string;
@@ -31,9 +32,8 @@ interface NoWorkspaceOauthClientCredsFlowProps {
  * NoWorkspaceOauthFlow first prompts user with a next button,
  * then launches a popup with the OAuth flow.
  */
-export function WorkspaceOauthClientCredsFlow({
-  provider, providerName,
-  consumerRef, consumerName, groupRef, groupName, explicitScopesRequired,
+export function NoWorkspaceOauthClientCredsFlow({
+  provider, consumerRef, consumerName, groupRef, groupName, explicitScopesRequired, providerName,
   selectedConnection, setSelectedConnection,
 }: NoWorkspaceOauthClientCredsFlowProps) {
   const { projectId } = useProject();
@@ -41,7 +41,7 @@ export function WorkspaceOauthClientCredsFlow({
   const [error, setError] = useState<string | null>(null);
 
   //  fetch OAuth callback URL from connection so that oath popup can be launched
-  const handleSubmit = async (creds: WorkspaceClientCredentialsCreds) => {
+  const handleSubmit = async (creds: ClientCredentialsCreds) => {
     setError(null);
     const req: GenerateConnectionRequest = {
       groupName,
@@ -49,7 +49,6 @@ export function WorkspaceOauthClientCredsFlow({
       consumerName,
       consumerRef,
       provider,
-      providerWorkspaceRef: creds.workspace,
       oauth2ClientCredentials: {
         clientId: creds.clientId,
         clientSecret: creds.clientSecret,
@@ -70,7 +69,7 @@ export function WorkspaceOauthClientCredsFlow({
 
   if (selectedConnection === null) {
     return (
-      <ClientCredentialsContent
+      <NoWorkspaceClientCredentialsContent
         providerName={providerName}
         handleSubmit={handleSubmit}
         error={error}
